requirements.txt
boto3==1.7.80
botocore==1.10.80
certifi==2018.8.13
chardet==3.0.4
dataclasses==0.6
dataclasses-json==0.0.11
docutils==0.14
idna==2.7
jmespath==0.9.3
PyJWT==1.6.4
PySocks==1.6.8
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
twilio==6.16.1
urllib3==1.23
